Add C++ decision tree class base that will use CUDA kernels for bootstrap sample, finding a node split, and splitting a node

+++ test.cu
@@ -7,7 +7,7 @@
 #include "CycleTimer.h"
 
 void test(double* input, int n, int p) {
-    int* device_input;
+    double* device_input;
 
     cudaMalloc((void **)&device_input, sizeof(double) * n * p);
     cudaMemcpy(device_input, input, sizeof(double) * n * p, cudaMemcpyHostToDevice);
